Recomendation system and fake generated info
Added the recomendation system using the customers orders and then orders from other customers who bought the same movies to recommend and using reviews to rank the movies within the recommendation system. Also used the  laravel fake info generator to seed the database with customers, orders and reviews so we could make the recomendations work.